+++ projects/testapp/include/comm_channels/nix_serial_port_bridge.h
@@ -19,8 +19,8 @@ public:
 
     virtual void stop();
     virtual void start();
-    virtual int receive(uint8_t* buffer, size_t len);
-    virtual void send(const uint8_t* buffer, size_t len);
+    virtual int receive(uint8_t* buffer, int len);
+    virtual void send(const uint8_t* buffer, int len);
 
     static void throw_system_error(const std::string &msg);
