+++ projects/testapp/src/comm_channels/udp_bridge.cpp
@@ -111,7 +111,7 @@ void UdpBridge::stop()
     }
 }
 
-int UdpBridge::receive(uint8_t* buffer, size_t len, int flags)
+int UdpBridge::receive(uint8_t* buffer, int len, int flags)
 { 
 #if defined(_WIN32)
     int size = sizeof(m_last_packet_addr);
@@ -141,7 +141,7 @@ int UdpBridge::receive(uint8_t* buffer, size_t len, int flags)
 }
 
 
- void UdpBridge::send(const uint8_t* buffer, size_t len, int flags)
+ void UdpBridge::send(const uint8_t* buffer, int len, int flags)
  {
     int ret = sendto(m_sock, reinterpret_cast<const char*>(buffer), len, flags, &m_last_packet_addr, sizeof(m_last_packet_addr));
